Asyn fully working using select. Possibly start new branch with asyncio? See what Xiaojun prefers Thursday?
